requirements.txt
flask>=2.3
gunicorn>=21.2
twilio>=8.0
requests>=2.28
schedule>=1.2
python-dotenv>=1.0
